Allow overriding GraphQL URL via env variable (#11)
* Allow overriding GraphQL URL via env variable

Allow developers to override the URL of the GraphQL server endpoint via
the `REACT_APP_GRAPHQL_SERVER_URL` environment variable. If it's not
set, default to `/graphql`.

One may use the `.env.local` file in the project root to set the
variable. See the Create React App [documentation][1] for more info.

[1]: https://create-react-app.dev/docs/adding-custom-environment-variables/

* Update readme with quick deployment guide