Remove 'waived' from valid test statuses
Being waived is seldom a useful distinct status – even when a test is
waived, we might still want to know what its original outcome was.
Therefore, remove the `waived` status and instead use a boolean flag
wherever we need to know the waive status, e.g. in the `TestStatusIcon`
component in this case.