Fix test selection dataflow
Make sure to store the requested focus test so that we can look it up
later, once all the tests have been loaded. This ensures that if you
select a test and refresh the page, it will expand automatically. This
was broken previously because the `useEffect()` hook only runs on first
render and when the query parameters change. When this happens at page
load, the `tests` variable is not yet populated becuse the query has not
finished yet. Therefore the correct test could not have been looked up
at this point in time.